Split BigUpdate into its own file
